+++ kraftwerk-api/src/main/java/fr/insee/kraftwerk/api/process/MainProcessing.java
@@ -117,7 +117,7 @@ public class MainProcessing {
 	}
 
 	/* Step 4 : Write output files */
-	private void outputFileWriter() {
+	private void outputFileWriter() throws KraftwerkException {
 		WriterSequence writerSequence = new WriterSequence();
 		writerSequence.writeOutputFiles(inDirectory, vtlBindings, userInputs.getModeInputsMap(),
 				userInputs.getMultimodeDatasetName(), metadataVariables, errors);