+++ kraftwerk-core/src/test/java/fr/insee/kraftwerk/core/metadata/GroupTest.java
@@ -4,16 +4,16 @@ import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-public class GroupTest {
+ class GroupTest {
 
     @Test
    // @SuppressWarnings({"null", "ConstantConditions"})
-    public void nullParentName() {
+    void nullParentName() {
         assertThrows(NullPointerException.class, () -> new Group("TEST", null));
     }
 
     @Test
-    public void emptyParentName() {
+    void emptyParentName() {
         assertThrows(IllegalArgumentException.class, () -> new Group("TEST", ""));
     }
 }