+++ kraftwerk-functional-tests/src/test/java/cucumber/functional_tests/ParadataDefinitions.java
@@ -78,7 +78,7 @@ public class ParadataDefinitions {
 		
 
 	}
-
+  
 	@Then("For UE {string} the collection date should be {string}")
 	public void check_collection_date(String identifier, String expectedCollectionDate) {
 		QuestionnaireData questionnaire = data.getQuestionnaires().stream().filter(